refactor: extract TableColGroup into column-groups/col-group.tsx
