+++ src/table/use-column-widths.tsx
@@ -6,8 +6,6 @@ import { useResizeObserver, useStableCallback } from '@cloudscape-design/compone
 import { getLogicalBoundingClientRect } from '@cloudscape-design/component-toolkit/internal';
 
 import { ColumnWidthStyle, setElementWidths } from './column-widths-utils';
-import { TableProps } from './interfaces';
-import { getColumnKey } from './utils';
 
 export const DEFAULT_COLUMN_WIDTH = 120;
 
@@ -266,42 +264,6 @@ export function ColumnWidthsProvider({
     >
       {children}
     </WidthsContext.Provider>
-  );
-}
-
-/*
- * Renders a <colgroup> with <col> elements for each column.
- * With table-layout:fixed, <col> widths control actual column widths,
- * which makes colspan headers automatically span the correct width.
- * Must be rendered inside ColumnWidthsProvider.
- */
-export function TableColGroup({
-  visibleColumnDefinitions,
-  hasSelection,
-  sticky = false,
-  selectionColumnId,
-}: {
-  visibleColumnDefinitions: ReadonlyArray<TableProps.ColumnDefinition<any>>;
-  hasSelection: boolean;
-  sticky?: boolean;
-  selectionColumnId?: PropertyKey;
-}) {
-  const { getColumnStyles, setCol } = useColumnWidths();
-  return (
-    <colgroup>
-      {hasSelection && (
-        <col
-          style={sticky && selectionColumnId ? { width: getColumnStyles(true, selectionColumnId).width } : undefined}
-        />
-      )}
-      {visibleColumnDefinitions.map((column, colIndex) => {
-        const columnId = getColumnKey(column, colIndex);
-        if (sticky) {
-          return <col key={String(columnId)} style={{ width: getColumnStyles(true, columnId).width }} />;
-        }
-        return <col key={columnId} data-column-id={String(columnId)} ref={node => setCol(columnId, node)} />;
-      })}
-    </colgroup>
   );
 }
